AND-6770 Move AssetReader to specific package
