+++ core/datasource/src/main/java/com/tangem/datasource/config/FeaturesLocalLoader.kt
@@ -2,7 +2,7 @@ package com.tangem.datasource.config
 
 import com.squareup.moshi.JsonAdapter
 import com.squareup.moshi.Moshi
-import com.tangem.datasource.asset.AssetReader
+import com.tangem.datasource.asset.reader.AssetReader
 import com.tangem.datasource.config.models.ConfigModel
 import com.tangem.datasource.config.models.ConfigValueModel
 import com.tangem.datasource.config.models.FeatureModel
