+++ core/datasource/src/main/java/com/tangem/datasource/di/AssetModule.kt
@@ -1,7 +1,7 @@
 package com.tangem.datasource.di
 
-import com.tangem.datasource.asset.AndroidAssetReader
-import com.tangem.datasource.asset.AssetReader
+import com.tangem.datasource.asset.reader.AndroidAssetReader
+import com.tangem.datasource.asset.reader.AssetReader
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
