+++ core/datasource/src/main/java/com/tangem/datasource/di/TestnetTokensStorageModule.kt
@@ -2,7 +2,7 @@ package com.tangem.datasource.di
 
 import com.squareup.moshi.Moshi
 import com.squareup.moshi.adapter
-import com.tangem.datasource.asset.AssetReader
+import com.tangem.datasource.asset.reader.AssetReader
 import com.tangem.datasource.local.testnet.DefaultTestnetTokensStorage
 import com.tangem.datasource.local.testnet.TestnetTokensStorage
 import dagger.Module
